Move TaskUtils to IntentUtils
